ingestion_service/requirements.txt
fastapi
uvicorn
fastapi
uvicorn
kafka-python
pydantic
